# docs/swagger.yaml
basePath: /api/v1
definitions:
  domain.Chapter:
    properties:
      chapterNumber:
        type: string
      createdAt:
        type: string
      id:
        type: string
      mangaID:
        type: string
      pages:
        items:
          type: string
        type: array
      title:
        description: Optional
        type: string
      updatedAt:
        type: string
    type: object
  domain.Comment:
    properties:
      chapterID:
        description: Nullable
        type: string
      content:
        type: string
      createdAt:
        type: string
      id:
        type: string
      mangaID:
        description: Nullable
        type: string
      updatedAt:
        type: string
      userID:
        type: string
    type: object
  domain.CommentWithUser:
    properties:
      chapterID:
        description: Nullable
        type: string
      content:
        type: string
      createdAt:
        type: string
      id:
        type: string
      mangaID:
        description: Nullable
        type: string
      updatedAt:
        type: string
      userID:
        type: string
      username:
        type: string
    type: object
  domain.Manga:
    properties:
      author:
        type: string
      coverImageURL:
        description: Use a pointer to handle NULL values
        type: string
      createdAt:
        type: string
      description:
        type: string
      genres:
        items:
          type: string
        type: array
      id:
        type: string
      status:
        $ref: '#/definitions/domain.MangaStatus'
      title:
        type: string
      updatedAt:
        type: string
    type: object
  domain.MangaStatus:
    enum:
    - ongoing
    - completed
    - hiatus
    - cancelled
    type: string
    x-enum-varnames:
    - StatusOngoing
    - StatusCompleted
    - StatusHiatus
    - StatusCancelled
...
  handler.createCommentRequest:
    properties:
      content:
        maxLength: 1000
        minLength: 1
        type: string
    required:
    - content
    type: object
...
  handler.userResponse:
    properties:
      email:
        type: string
      id:
        type: string
      username:
        type: string
    type: object
  repository.ToggleFavoriteResult:
    properties:
      isFavorited:
        description: True if the manga is now a favorite, false if it was removed.
        type: boolean
    type: object
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is a production-ready HTTP API in Go for a manga reading platform.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Manga-Dex-Style API
  version: "1.0"
paths:
...
  /chapters/{id}/comments:
    get:
      description: Retrieves a paginated list of comments for a specific chapter.
      parameters:
      - description: Chapter ID
        in: path
        name: id
        required: true
        type: string
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 20
        description: Items per page
        in: query
        name: per_page
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/domain.CommentWithUser'
            type: array
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: List chapter comments
      tags:
      - Social
    post:
      consumes:
      - application/json
      description: Adds a new comment to a specific chapter.
      parameters:
      - description: Chapter ID
        in: path
        name: id
        required: true
        type: string
      - description: Comment Content
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.createCommentRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/domain.Comment'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Post a comment on a chapter
      tags:
      - Social
...
  /chapters/{id}/progress:
    post:
      description: Marks a chapter as read for the current user.
      parameters:
      - description: Chapter ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Mark chapter as read
      tags:
      - Social
...
  /manga/{id}/comments:
    get:
      description: Retrieves a paginated list of comments for a specific manga.
      parameters:
      - description: Manga ID
        in: path
        name: id
        required: true
        type: string
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 20
        description: Items per page
        in: query
        name: per_page
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/domain.CommentWithUser'
            type: array
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: List manga comments
      tags:
      - Social
    post:
      consumes:
      - application/json
      description: Adds a new comment to a specific manga.
      parameters:
      - description: Manga ID
        in: path
        name: id
        required: true
        type: string
      - description: Comment Content
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.createCommentRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/domain.Comment'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Post a comment on a manga
      tags:
      - Social
  /manga/{id}/favorite:
    post:
      description: Adds or removes a manga from the current user's favorites list.
      parameters:
      - description: Manga ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/repository.ToggleFavoriteResult'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Toggle manga favorite status
      tags:
      - Social
...
  /users/me:
    get:
      description: Retrieves the profile information for the currently authenticated
        user.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.userResponse'
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get current user's profile
      tags:
      - Users
  /users/me/favorites:
    get:
      description: Retrieves a paginated list of the current user's favorite manga.
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 20
        description: Items per page
        in: query
        name: per_page
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/domain.Manga'
            type: array
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: List user's favorite manga
      tags:
      - Social
  /users/me/progress:
    get:
      description: Retrieves a list of all chapters marked as read by the current
        user.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/domain.Chapter'
            type: array
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: List user's read chapters
      tags:
      - Social
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and a JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"